In progress Scraper
Scraper currently finds item names and price just fine, but runs into issues with the rating. Apparently people have had issues locating li classes inside ul, so I'm working on finding a solution to that, because at the moment, this code cannot locate the stars classes to construct the ratings.